pyproject.toml
[project]
name = "rimscode-website"
version = "0.1.0"
description = "Builder module for the RIMS-Code website scheme database."
authors = [
    { name = "Reto Trappitsch", email = "reto@galactic-forensics.space" }
]
dependencies = [
    "mkdocs>=1.5.3",
    "mkdocs-material>=9.5.4",
    "numpy>=1.26.3",
    "pyyaml>=6.0.1",
]
readme = "README.md"
requires-python = ">= 3.11"
license = { text = "MIT" }

[project.scripts]
website = "rimscode_website:website"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = [
    "mkdocs>=1.5.3",
    "mkdocs-material>=9.5.4",
    "ruff>=0.1.14",
    "pytest>=7.4.4",
    "pytest-mock>=3.12.0",
]

[tool.rye.scripts]
serve = "mkdocs serve"

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/rimscode_website"]
